fix: 좋아요 status 가 true 인 것만 조회되도록 변경
