+++ apps/api/src/diary/controller/diary.controller.ts
@@ -18,9 +18,8 @@ import {
   GetMyDiary,
   UpdateDiary,
   CreateDiaryImage,
-  GenerateDiaryLike,
 } from './decorator/diary.controller.decorator';
-import { DiaryLikeService, DiaryService } from '../service';
+import { DiaryService } from '../service';
 import {
   CreateDiaryRequestDto,
   GetDiaryListQueryRequestDto,
@@ -29,16 +28,12 @@ import {
   UpdateDiaryRequestDto,
   GetMyDiaryListRequestDto,
   CreateDiaryImageRequestDto,
-  GenerateDiaryLikeRequestDto,
 } from '../dto/requests';
 import { GetDiaryListResponseDto, GetDiaryResponseDto } from '../dto/responses';
 
 @Controller()
 export class DiaryController {
-  constructor(
-    private readonly diaryService: DiaryService,
-    private readonly diaryLikeService: DiaryLikeService,
-  ) {}
+  constructor(private readonly diaryService: DiaryService) {}
 
   @CreateDiary()
   @FileListRequest([{ name: 'diaryImageFile', maxCount: 5 }])
@@ -118,16 +113,6 @@ export class DiaryController {
       throw new FileRequiredException('diaryImageFile');
     }
     await this.diaryService.createDiaryImage(dto, diaryImageFile, userId);
-
-    return null;
-  }
-
-  @GenerateDiaryLike()
-  async generateDiaryLike(
-    @Jwt() { userId }: JwtRequestDto,
-    @Body() dto: GenerateDiaryLikeRequestDto,
-  ): Promise<null> {
-    await this.diaryLikeService.createOrUpdateDiaryLikes(dto, userId);
 
     return null;
   }
