+++ apps/api/src/diary/controller/index.ts
@@ -1,1 +1,2 @@
 export * from './diary.controller';
+export * from './diary-like.controller';
