+++ apps/api/src/diary/diary.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 
 import { UploadFileModule } from '@app/upload-file';
 
-import { DiaryController } from './controller';
+import { DiaryController, DiaryLikeController } from './controller';
 import {
   DiaryServiceImpl,
   DiaryService,
@@ -17,7 +17,7 @@ import {
 
 @Module({
   imports: [UploadFileModule],
-  controllers: [DiaryController],
+  controllers: [DiaryController, DiaryLikeController],
   providers: [
     {
       provide: DiaryService,
